+++ TFG/src/main/java/com/project/controllers/ClientController.java
@@ -1,9 +1,7 @@
 package com.project.controllers;
 
 import java.util.HashMap;
-import java.util.List;
 import java.util.Map;
-
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.dao.DataAccessException;
 import org.springframework.http.HttpStatus;
@@ -11,10 +9,7 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.CrossOrigin;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RestController;
-
-
 import com.project.domain.Client;
 import com.project.services.ClientService;
 
@@ -30,7 +25,7 @@ public class ClientController {
 	// -------------------------- Show ----------------------------------
 	@CrossOrigin
 	@RequestMapping("/show/{id}")
-	public ResponseEntity<?> show(@PathVariable Long id) {
+	public ResponseEntity<?> show(@PathVariable int id) {
 		Client client = null;
 		Map<String, Object> response = new HashMap<>();
 		try {
