+++ TFG/src/main/java/com/project/controllers/FactureCompanyController.java
@@ -36,7 +36,7 @@ public class FactureCompanyController {
 	// ---------------------------- Show facture----------------------------------------------------------
 	@CrossOrigin
 	@RequestMapping("/show/{id}")
-	public ResponseEntity<?> show(@PathVariable Long id) {
+	public ResponseEntity<?> show(@PathVariable int id) {
 		Facture facture = null;
 		Map<String, Object> response = new HashMap<>();
 		try {
