+++ TFG/src/main/java/com/project/domain/Role.java
@@ -18,19 +18,19 @@ public class Role implements Serializable {
 
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private Long id;
+	private int id;
 
 	@Column(unique = true, length = 20)
 	private String nombre;
 
-	public Long getId() {
+	public int getId() {
 		return id;
 	}
 
-	public void setId(Long id) {
+	public void setId(int id) {
 		this.id = id;
 	}
-
+ 
 	public String getNombre() {
 		return nombre;
 	}
