+++ TFG/src/main/java/com/project/repositories/BasketRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import com.project.domain.Basket;
 
 @Repository
-public interface BasketRepository extends JpaRepository<Basket, Long> {
+public interface BasketRepository extends JpaRepository<Basket, Integer> {
 
 }
