+++ TFG/src/main/java/com/project/repositories/ClientRepository.java
@@ -2,7 +2,7 @@ package com.project.repositories;
 
 
 import org.springframework.data.jpa.repository.JpaRepository;
-
+import org.springframework.data.jpa.repository.Query;
 import org.springframework.stereotype.Repository;
 
 
@@ -10,6 +10,7 @@ import com.project.domain.Client;
 
 
 @Repository
-public interface ClientRepository extends JpaRepository<Client, Long> {
+public interface ClientRepository extends JpaRepository<Client, Integer> {
 	
+
 }
