+++ TFG/src/main/java/com/project/repositories/ConfigurationRepository.java
@@ -10,7 +10,7 @@ import org.springframework.stereotype.Repository;
 import com.project.domain.Configuration;
 
 @Repository
-public interface ConfigurationRepository extends JpaRepository<Configuration, Long> {
+public interface ConfigurationRepository extends JpaRepository<Configuration, Integer> {
 	
 
 }
