+++ TFG/src/main/java/com/project/services/ClaimService.java
@@ -1,14 +1,10 @@
 package com.project.services;
 
 import java.util.List;
-
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-
-import com.project.domain.Admin;
 import com.project.domain.Claim;
-
 import com.project.repositories.ClaimRepository;
 
 @Service
@@ -26,14 +22,14 @@ public class ClaimService {
 	// ----------------------------------------List claims by
 	// facture------------------------------------------------------------------
 	@Transactional(readOnly = true)
-	public List<Claim> findClaimByFacture(Long clientId) {
+	public List<Claim> findClaimByFacture(int clientId) {
 		return (List<Claim>) claimRepository.findClaimByFacture(clientId);
 	}
 
 	// -----------------------------------------Show claim
 	// -----------------------------------------------------------------------------
 	@Transactional(readOnly = true)
-	public Claim findById(Long id) {
+	public Claim findById(int id) {
 		return claimRepository.findById(id).orElse(null);
 
 	}
